pyproject.toml
[project]
name = "devops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "joblib",
    "scikit-learn",
    "pytest",
    "azureml-core",
    "azureml-sdk",
    "numpy",
    "pillow",
    "setuptools",
    "dotenv>=0.9.9",
]
